UniFFI-export the offline presentation module
Exposes core/src/presentation/{request,verifiable_presentation,
offline_verifier}.rs to Android for the first time: minting/encoding/
decoding a verifier's PresentationRequest, wrapping a stored
credential as an EnvelopedVerifiableCredential for presenting,
building and assembling a presentation JWS (holder side), and
verifying a presentation shown to this device (verifier side). All
three modules were already fully ported and tested — this pass is
pure FFI wiring, no new algorithms.

Follows this crate's established precedent throughout: derive
uniffi::Record/Enum/Error directly on domain types in their home
module (PresentationRequest, PresentationCredentialSource,
EnvelopedVerifiableCredential, VerificationFailure,
VerificationCaveat, RevocationStatus, NotCheckedReason,
OfflineIssuerTrustSnapshot, DisclosedClaim, RevocationSnapshotInfo,
and both error enums), and mirror only where a field shape isn't
UniFFI-representable - one new FfiVerifiedPresentation record in
ffi.rs, converting VerifiedPresentation's three DateTime<Utc> fields
to Unix-seconds i64, matching the existing
ConvenienceStorePickupBarcode precedent exactly.

Two small additions beyond pure derives, both mechanical: changed
VerificationFailure::PresentationCarriesMultipleCredentials's `count`
from usize (not UniFFI-representable) to u32, and added
VerifiablePresentationError::InvalidTimestamp for the one new failure
mode an integer-timestamp FFI boundary introduces that DateTime<Utc>
never had.

Also exposes presentation_subject_identifier/
presentation_moica_subject_identifier - not in the original scope
list, but trivial pure getters already in the ported module, and
necessary for a caller to know which stored credential's subject
matches a request's credential_source before presenting at all.

Real BLE/QR transport, and the Android UI to drive any of this, stay
out of scope - the actual iOS transport turned out to be a genuine
multi-frame QR carousel racing a BLE transfer, both with their own
framing protocols, worth porting to core on their own rather than
alongside this. Tracked as follow-ups.

354 -> 378 tests (24 new), full suite passing. clippy/fmt clean.
./gradlew :wallet:assembleDebug confirms the generated Kotlin bindings
for the new FFI surface actually compile.

+++ core/src/presentation/offline_verifier.rs
@@ -50,7 +50,7 @@ use crate::twdiw::onchain;
 /// say on its own. Every one is attached to a presentation that passed
 /// every check - these exist because a bare 「驗證通過」 lets a verifier
 /// read guarantees into the result that this design does not provide.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, uniffi::Enum)]
 pub enum VerificationCaveat {
     NoNetworkQuery,
     RevocationNotChecked,
@@ -66,7 +66,7 @@ pub enum VerificationCaveat {
     NotBoundToThisVerifier,
 }
 
-#[derive(Debug, Clone, PartialEq, thiserror::Error)]
+#[derive(Debug, Clone, PartialEq, thiserror::Error, uniffi::Error)]
 pub enum VerificationFailure {
     // Structure
     #[error("presentation is not a JWS")]
@@ -94,7 +94,7 @@ pub enum VerificationFailure {
     #[error("credential missing")]
     CredentialMissing,
     #[error("presentation carries multiple credentials: {count}")]
-    PresentationCarriesMultipleCredentials { count: usize },
+    PresentationCarriesMultipleCredentials { count: u32 },
     #[error("credential not enveloped")]
     CredentialNotEnveloped,
     #[error("credential is not a JWS")]
@@ -155,7 +155,7 @@ pub enum VerificationFailure {
     DeviceClockPrecedesCertificate { valid_from: i64 },
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq, uniffi::Record)]
 pub struct DisclosedClaim {
     pub term: String,
     pub value: String,
@@ -209,14 +209,14 @@ impl VerificationOutcome {
 
 // MARK: - Revocation (shape only - see module docs)
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq, uniffi::Enum)]
 pub enum RevocationStatus {
     Revoked { snapshot: RevocationSnapshotInfo },
     NotRevokedInThisSnapshot { snapshot: RevocationSnapshotInfo },
     NotChecked { reason: NotCheckedReason },
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, uniffi::Enum)]
 pub enum NotCheckedReason {
     SnapshotUnavailable,
     SnapshotUnusable,
@@ -224,7 +224,7 @@ pub enum NotCheckedReason {
     NoCertificateToCheck,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq, uniffi::Record)]
 pub struct RevocationSnapshotInfo {
     pub root: String,
     /// `YYYYMMDDHH`.
@@ -288,7 +288,7 @@ pub fn caveat_for_revocation_status(
 /// Reading/writing this snapshot to disk stays native
 /// (`docs/2026-09-05-decisions-and-roadmap.md`); this module only compares
 /// one already-looked-up snapshot against a credential's issuer.
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq, uniffi::Record)]
 pub struct OfflineIssuerTrustSnapshot {
     pub issuer_did: String,
     pub display_name: String,
@@ -617,7 +617,7 @@ fn enveloped_credential(
     if entries.len() != 1 {
         return Err(
             VerificationFailure::PresentationCarriesMultipleCredentials {
-                count: entries.len(),
+                count: entries.len() as u32,
             },
         );
     }
